+++ core/test/test_cambi_simd.c
@@ -1,5 +1,6 @@
 /**
  *
+ *  Copyright 2016-2026 Netflix, Inc.
  *  Copyright 2026 Lusoris
  *
  *     Licensed under the BSD+Patent License (the "License");
@@ -39,6 +40,12 @@
  */
 
 #include <stddef.h>
+
+/* NOLINTBEGIN(modernize-use-nullptr): C translation unit. The fork builds C as
+ * C23, where clang-tidy also proposes the `nullptr` keyword, but MSVC's
+ * documented /std:clatest C23 feature set does not include `nullptr` while the
+ * required Windows build compiles this TU with cl.exe, and this file mirrors
+ * the C spelling of the surface it exercises. ADR-1138. */
 #include <stdint.h>
 #include <stdio.h>
 #include <string.h>
@@ -84,7 +91,7 @@ static void calculate_c_values_row_scalar(float *c_values, const uint16_t *histo
 
     const uint16_t *image_row = &image[row * stride];
     const uint16_t *mask_row = &mask[row * stride];
-    float *c_row = &c_values[row * width];
+    float *c_row = &c_values[(size_t)row * (size_t)width];
 
     for (int col = 0; col < width; col++) {
         if (!mask_row[col]) {
@@ -381,3 +388,5 @@ char *run_tests(void)
 
     return NULL;
 }
+
+/* NOLINTEND(modernize-use-nullptr) */
